# Support different data type between input and output