<!-- Demo: https://github.com/PaddlePaddle/Paddle/pull/24810 -->
### PR types
<!-- One of [ New features | Bug fixes | Function optimization | Performance optimization | Breaking changes | Others ] -->
Performance optimization
### PR changes
<!-- One of [ OPs | APIs | Docs | Others ] -->
OPs
### Describe
<!-- Describe what this PR does -->
elementwise 实现模版支持输入和输出的数据类型不同，并适配 abs 验证，性能数据如下：
|config |dt|paddle old|paddle new  |pro  |pytorch  |pro  | 
|---|---|---|---|---|---|---|
|[16, 128, 257, 257] |fp16|852.23us|668.53us | 27.48% |750.56us| 12.27% | 
|[16, 128, 257, 257]|fp32|1.3771ms|1.3109ms|5.05% |1.3115ms | 0.05% | 
优化后，性能已经打平甚至优于 pytorch